functions/python/requirements.txt
functions-framework==3.*
pdfplumber==0.11.*
pandas==2.*
openpyxl==3.*
xlrd==2.*
python-dateutil==2.8.*
